Fix for building with gcc 3.3 from Andrew Cottrell.

+++ src/libemu/src/emu387.cc
@@ -272,7 +272,7 @@ static void emu_printall()
           (long)(r->sigl & 0xFFFF),
           r->exp - EXP_BIAS + 1);
     }
-    eprintf("%s\r\n", tag_desc[r->tag]);
+    eprintf("%s\r\n", tag_desc[(int)r->tag]);
   }
 }
